Fixed Some Warnings from the Pull

+++ ChronoLoop/ChronoLoop/Objects/BaseObject.cpp
@@ -158,7 +158,7 @@ unsigned int BaseObject::AddComponent(Component * _comp) {
 	}
 	_comp->mObject = this;
 	mComponents[_comp->GetType()].push_back(_comp);
-	_comp->mComponentNum = mComponents.size();	//sets the component number for knowing the position in the bitset
+	_comp->mComponentNum = (unsigned short)mComponents.size();	//sets the component number for knowing the position in the bitset
 	return (unsigned int)mComponents[_comp->GetType()].size();
 }
